Float to double

+++ src/core/datatypes.h
@@ -8,7 +8,7 @@ typedef uint32_t vertex_id_t;
 
 typedef uint32_t graph_size_t;
 
-typedef float state_t;
+typedef double state_t;
 
 typedef uint32_t task_id_t;
